# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NOC API",
        "description": "ILRMS/NOC land sale workflow API - ADC, DC, CO, LRA, ASST, JS modules.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "NOC API server"
        }
    ],
    "paths": {
        "/noc/applications": {
            "post": {
                "tags": [
                    "NOC Applications"
                ],
                "summary": "Get NOC applications for a posting",
                "operationId": "92d2c7566b1306ac8351511ed80dd72b",
                "parameters": [
                    {
                        "name": "posting_uuid",
                        "in": "query",
                        "description": "UUID of the posting",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "nullable": true
                        }
                    },
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "Filter applications by status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "inbox",
                                "pending",
                                "approved",
                                "returned",
                                "rejected",
                                "query_sent",
                                "query_responded"
                            ]
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search applications by application number",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "nullable": true
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of records per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10,
                            "maximum": 100,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "action_filter",
                        "in": "query",
                        "description": "Filter by action code",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "nullable": true
                        }
                    },
                    {
                        "name": "distcode",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "nullable": true
                        }
                    },
                    {
                        "name": "subcode",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "nullable": true
                        }
                    },
                    {
                        "name": "circode",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "nullable": true
                        }
                    },
                    {
                        "name": "mouzacode",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "nullable": true
                        }
                    },
                    {
                        "name": "lotno",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "nullable": true
                        }
                    },
                    {
                        "name": "villcode",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "nullable": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of NOC applications"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/land/schedule/details": {
            "post": {
                "tags": [
                    "Applications Details"
                ],
                "summary": "Get Land Schedule Details",
                "description": "Fetches land schedule details based on the provided application number.",
                "operationId": "27dbd21a5ae071694b0c6a55018568fc",
                "parameters": [
                    {
                        "name": "app_no",
                        "in": "query",
                        "description": "Unique application number used to retrieve the land schedule details.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "nullable": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Land Schedule Details Retrieved Successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/application/view": {
            "post": {
                "tags": [
                    "Applications Details"
                ],
                "summary": "View application details",
                "description": "Fetches application details using the application number. The API retrieves applicant information, associated land sale details, and the latest tracking status of the application.",
                "operationId": "54cf90f634d1b0277ccebf757ece42dd",
                "parameters": [
                    {
                        "name": "appno",
                        "in": "query",
                        "description": "Unique application number used to retrieve the application details.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "nullable": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Applications Details Retrieved Successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT access token.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "NOC Applications",
            "description": "NOC Applications"
        },
        {
            "name": "Applications Details",
            "description": "Applications Details"
        }
    ]
}